+++ packages/admin-api-client/src/types/index.ts
@@ -8,7 +8,9 @@ import type {
 import type {
   ApiResponseBody as _ApiResponseBody,
   ApiClientConfig as _ApiClientConfig,
-  
+  Protocol as _Protocol,
+  Url as _Url,
+  BaseUrl as _BaseUrl
 } from './AdminApiClientEntities';
 
 import type { EntityId as _EntityId } from "./Common";
@@ -17,6 +19,9 @@ export namespace ApiClient {
   export type ResponseBody = _ApiResponseBody;
   export type Config = _ApiClientConfig;
   export type EntityId = _EntityId;
+  export type Protocol = _Protocol;
+  export type BaseUrl = _BaseUrl;
+  export type Url = _Url;
 
   export type Interface = _AdminApiClientInterface;
   export type Constructor = _AdminApiClientConstructor;