+++ packages/cypress-commands/src/index.ts
@@ -1,8 +1,13 @@
-export * from "./types";
+export * from "./types/index";
+
 export { AdminApiClient } from "./AdminApiClient";
 
 import { register as registerer } from "./register";
 
-export function register() {
+/**
+* Register Mocks Server Cypress commands
+* @example register();
+*/
+export function register(): void {
   registerer(Cypress, cy);
 }