+++ packages/cypress-commands/CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 ### Removed
 
+## [unreleased]
+
+### Changed
+- feat: Remove legacy mocksConfigAdminApiClient command. mocksConfigClient has to be used instead
+- refactor: Change exported types. Use a namespace
+
 ## [6.1.1] - 2022-09-14
 
 ### Changed